Don't mess around with instance_exec.
Although it looks more DSLish with instance_exec it is dangerous because
it changes self and the user may not expect that. Passing the context as
block argument and forcing the user to explicitly send method to it is may be
less DSLish but way simpler and less error prone.